Moved legend code into its own file
